# doc/openapi.yaml
openapi: 3.0.3
info:
  title: Materna
  description: API REST do projeto Materna para gerenciamento de doações de leite
    materno, desenvolvida como parte do TCC
  termsOfService: http://swagger.io/terms/
  contact:
    email: brunohaddadalves75@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.4.26
...
servers:
- url: http://localhost:9090
paths:
  /v1/materna/doadora:
    post:
      tags:
      - Doadora
      summary: Adiciona uma nova doadora
      description: Endpoint para enviar os dados de uma doadora a ser cadastrada
      requestBody:
        description: JSON contendo os dados de uma doadora
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Doadora'
      responses:
        '201':
          description: Doadora cadastrada com sucesso
        '400':
          description: Invalid input
        '415':
          description: Invalid Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    get:
      tags:
      - Doadora
      summary: Lista todas as doadoras
      description: Listagem de todas as doadoras registradas no sistema
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Doadora'
        '404':
          description: Not Found
        '500':
          description: Error Internal Server (Controller / Model)
  /v1/materna/doadora/{id}:
    get:
      tags:
      - Doadora
      summary: Busca uma doadora pela ID
      description: Retorna uma doadora pela ID
      parameters:
      - name: id
        in: path
        description: ID da doadora
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Doadora'
        '400':
          description: Invalid ID supplied
        '404':
          description: Not Found
        '500':
          description: Error Internal Server (Controller / Model)
    put:
      tags:
      - Doadora
      summary: Atualiza uma doadora pela ID
      description: Buscar e atualizar uma doadora pela ID
      parameters:
      - name: id
        in: path
        description: ID da doadora
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: Dados para atualizar uma doadora
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Doadora'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Doadora not found
        '415':
          description: Invalid Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    delete:
      tags:
      - Doadora
      summary: Deletar uma doadora existente
      description: Excluir uma doadora
      parameters:
      - name: id
        in: path
        description: ID da doadora
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Doadora not found
        '500':
          description: Error Internal Server (Controller / Model)
...
components:
  schemas:
    Doadora:
      type: object
      required:
      - nome
      - cpf
      - data_nascimento
      - email
      - senha
      - telefone
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Maria Silva
        cpf:
          type: string
          example: 123.456.789-00
        foto:
          type: string
          example: https://exemplo.com/foto.jpg
        data_nascimento:
          type: string
          format: date
          example: '1995-04-20'
        email:
          type: string
          example: maria@email.com
        senha:
          type: string
          example: maria123
        telefone:
          type: string
          example: '11949491088'
        cep:
          type: string
          example: 92330-036
        estado:
          type: string
          example: SP
        cidade:
          type: string
          example: Itapevi
        endereco:
          type: string
          example: Av.Castelo Branco
        numero:
          type: integer
          example: 11
        complemento:
          type: string
          example: 9A
...